Move DOM targeting helpers into dom
Constraint: Element discovery, AX resolution, shadow DOM traversal, and DOM delta feedback form a DOM targeting boundary.\nRejected: Keeping these modules under src/utils | they are browser DOM behavior, not cross-cutting leaf helpers.\nRejected: Moving puppeteer target helpers in this PR | getTargetId is shared by CDP/session/chrome paths and needs its own classification.\nConfidence: high\nScope-risk: moderate\nDirective: Put new DOM targeting and traversal code under src/dom, not src/utils.\nTested: npx jest --config jest.ci.config.js --runInBand --silent --runTestsByPath tests/dom/element-affordance.test.ts tests/dom/ax-element-resolver.test.ts tests/dom/shadow-dom-discovery.test.ts tests/dom/shadow-dom-query.test.ts tests/dom/shadow-dom-utils.test.ts tests/tools/find.test.ts tests/tools/act.test.ts tests/tools/interact.coordinate.test.ts tests/tools/interact-opened-tabs.test.ts tests/tools/inspect-metrics.test.ts tests/tools/snapshot-refs.test.ts tests/vision/vision-find.test.ts tests/headed/headed-tool-interop.test.ts tests/headed/headed-fallback-manager.test.ts\nTested: npx jest --config jest.ci.config.js --runInBand --silent --runTestsByPath tests/utils/ralph/ralph-engine.test.ts tests/dom/element-affordance.test.ts tests/dom/ax-element-resolver.test.ts tests/dom/shadow-dom-discovery.test.ts tests/dom/shadow-dom-query.test.ts tests/dom/shadow-dom-utils.test.ts tests/tools/find.test.ts tests/tools/act.test.ts tests/tools/interact.coordinate.test.ts\nTested: npm run build\nTested: npm run lint\nTested: npm run lint:tier\nTested: npm run lint:repo-structure\nTested: git diff --check\nNot-tested: Full Jest suite.
