+++ docs/dev/project-structure.md
@@ -71,6 +71,7 @@ belongs in `skills/` and `commands/`; do not fork equivalent copies under
 | `src/mcp/` | MCP server implementation, protocol ingress helpers, session-init policy, and MCP output accounting. `src/mcp-server.ts` is a compatibility re-export for existing imports. |
 | `src/transports/` | MCP/HTTP transport surfaces and protocol adapters. |
 | `src/chrome/`, `src/cdp/`, `src/browser-state/` | Chrome process, PID/guardian cleanup, controller ownership, CDP, and browser state integration. |
+| `src/dom/` | DOM serialization, element discovery, AX resolution, shadow DOM traversal, and DOM-change feedback. |
 | `src/session/` | session ownership, lease, snapshot, and lifecycle coordination. `src/session-manager.ts` is a compatibility re-export for existing imports; new code should import from `src/session/manager` or the nearest package entrypoint. |
 | `src/router/` | routing browser/tool commands to the correct target or tab. |
 | `src/pilot/` | higher-level agent runtime features: skills, handoff, voting, credentials, curator, and automation runtime. |
