+++ src/dom/element-affordance.ts
