+++ src/tools/fill-form.ts
@@ -10,12 +10,12 @@ import { TOOL_ANNOTATIONS } from '../types/tool-annotations';
 import { getSessionManager } from '../session-manager';
 import { getRefIdManager, formatStaleRefError, makeStaleRefError } from '../utils/ref-id-manager';
 import { DEFAULT_DOM_SETTLE_DELAY_MS, DEFAULT_FORM_SUBMIT_SETTLE_MS } from '../config/defaults';
-import { withDomDelta } from '../utils/dom-delta';
+import { withDomDelta } from '../dom/dom-delta';
 import { withTimeout } from '../utils/with-timeout';
-import { discoverFormFields, FormField, FORM_FIELD_TAG } from '../utils/element-discovery';
-import { resolveElementsByAXTree, invalidateAXCache } from '../utils/ax-element-resolver';
+import { discoverFormFields, FormField, FORM_FIELD_TAG } from '../dom/element-discovery';
+import { resolveElementsByAXTree, invalidateAXCache } from '../dom/ax-element-resolver';
 import { getTargetId } from '../utils/puppeteer-helpers';
-import { normalizeQuery } from '../utils/element-finder';
+import { normalizeQuery } from '../dom/element-finder';
 import { humanType, humanMouseMove } from '../stealth/human-behavior';
 import { isPilotEnabled } from '../harness/flags';
 import { detectLoginOutcome, LoginDetectResult } from './login-detector';
