+++ src/tools/interact.ts
@@ -9,13 +9,13 @@ import { MCPServer } from '../mcp-server';
 import { MCPToolDefinition, MCPResult, ToolHandler, ToolContext, hasBudget, throwIfAborted } from '../types/mcp';
 import { getSessionManager } from '../session-manager';
 import { getRefIdManager } from '../utils/ref-id-manager';
-import { withDomDelta } from '../utils/dom-delta';
+import { withDomDelta } from '../dom/dom-delta';
 import { DEFAULT_DOM_SETTLE_DELAY_MS, DEFAULT_SCREENSHOT_RACE_TIMEOUT_MS, DEFAULT_SCREENSHOT_TIMEOUT_MS } from '../config/defaults';
-import { FoundElement, normalizeQuery, scoreElement, tokenizeQuery } from '../utils/element-finder';
-import { discoverElements, getTaggedElementRect, cleanupTags, DISCOVERY_TAG } from '../utils/element-discovery';
+import { FoundElement, normalizeQuery, scoreElement, tokenizeQuery } from '../dom/element-finder';
+import { discoverElements, getTaggedElementRect, cleanupTags, DISCOVERY_TAG } from '../dom/element-discovery';
 import { withTimeout } from '../utils/with-timeout';
 import { makeImageContent } from '../utils/image-mime';
-import { resolveElementsByAXTree, invalidateAXCache, MATCH_LEVEL_LABELS } from '../utils/ax-element-resolver';
+import { resolveElementsByAXTree, invalidateAXCache, MATCH_LEVEL_LABELS } from '../dom/ax-element-resolver';
 import { getTargetId } from '../utils/puppeteer-helpers';
 import { classifyOutcome, formatOutcomeLine } from '../utils/ralph/outcome-classifier';
 import { getCircuitBreaker } from '../utils/ralph/circuit-breaker';
