+++ src/tools/query-dom.ts
@@ -10,7 +10,7 @@ import { MCPToolDefinition, MCPResult, ToolHandler, ToolContext, hasBudget } fro
 import { TOOL_ANNOTATIONS } from '../types/tool-annotations';
 import { getSessionManager } from '../session-manager';
 import { withTimeout } from '../utils/with-timeout';
-import { getAllShadowRoots, querySelectorInShadowRoots } from '../utils/shadow-dom';
+import { getAllShadowRoots, querySelectorInShadowRoots } from '../dom/shadow-dom';
 import { isSnapshotCacheEnabled } from '../utils/snapshot-cache-helper';
 import { getCurrentLoaderId, mintNodeRefSync } from '../core/perception/node-ref';
 import { decodeCursor, encodeCursor, paginate } from '../utils/paginate';
